Bugfix: Online game doesn't handle repeated letters correctly
The online game marks the first instance of a letter as correct but not the second. The tool blindly added the letter to both the good/known dictionary AND the bad letter list. I resolved this by removing it from the list when we go to solve it.
I also prematurely optimized the list of bad letters to contain only unique characters, no repeats allowed.